+++ tests/resources/tf/link_default_single.tf
@@ -1,0 +1,7 @@
+resource "parent" "parent1" {
+    name="parent1"
+    toChild="child1"
+}
+resource "child" "child1" {
+    name="child1"
+}