+++ changelog.md
@@ -13,5 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setup default plugin structure
 - Set default metadata parser and validator
 - Add metadata for aws
+- Add parser for terraform files
 
 [Unreleased]: https://github.com/ditrit/terrator-plugin/blob/main/changelog.md